requirements.txt
openai
python-dotenv
transformers
accelerate
peft
bitsandbytes
datasets
fire
huggingface_hub
torch
torchvision
trackio
wandb